* origin/master: (427 commits)
  refactor: make PSQLState enum consts for integrity constraint violations (#1699)
  [maven-release-plugin] prepare for next development iteration
  [maven-release-plugin] prepare release REL42.2.10
  pass gpg key through arguments
  add passphrase to release mvn task
  chore: update signing key
  Metadata queries improvment (#1694)
  WIP release notes for 42.2.10 (#1688)
  chore(deps): bump checkstyle from 8.28 to 8.29 in /pgjdbc (#1691)
  Cleanup PGProperty, sort values, and add some missing to docs (#1686)
  fix: Fixes issue #1592 where one thread is reading the copy and another thread closes the connection (#1594)
  Fixing LocalTime rounding (losing precision) (#1570)
  sync error message value with tested value (#1664)
  add DatabaseMetaDataCacheTest to test suite to run it (#1685)
  Fix Network Performance of PgDatabaseMetaData.getTypeInfo() method (#1668)
  fix: Issue #1680 updating a boolean field requires special handling to set it to t or f instead of true or false (#1682)
  fix testSetNetworkTimeoutEnforcement test failure (#1681)
  minor: fix checkstyle violation of unused import (#1683)
  fix: #1677 NumberFormatException when fetching PGInterval with small value (#1678)
  fix: actually use milliseconds instead of microseconds for timeouts (#1653)
  ...

+++ pgjdbc/src/main/java/org/postgresql/core/ResultHandlerBase.java
@@ -30,7 +30,7 @@ public class ResultHandlerBase implements ResultHandler {
   }
 
   @Override
-  public void handleCommandStatus(String status, int updateCount, long insertOID) {
+  public void handleCommandStatus(String status, long updateCount, long insertOID) {
   }
 
   @Override
